Missing values imputation wrapper

Missing values imputation wrapper

impute_missing(dataset, method, exclude = NULL, ...)

Arguments

dataset
we want to impute missing values on

method
selected method of missing values imputation

exclude
character. Vector of attributes to exclude from the missing values treatment

...
Further arguments for method

Value

The treated dataset (either with noisy instances replaced or erased)

Examples

library("smartdata")
data(africa, package = "Amelia")
data(nhanes, package = "mice")
data(ozone,  package = "missMDA")
data(vnf,    package = "missMDA")
data(orange, package = "missMDA")
data(sleep,  package = "VIM")

super_nhanes <- impute_missing(nhanes, "gibbs_sampling")
super_nhanes <- impute_missing(nhanes, "gibbs_sampling", exclude = "chl")
# Use a different method for every column
impute_methods <- c("pmm", "midastouch", "norm_nob", "norm_boot")
super_nhanes <- impute_missing(nhanes, "gibbs_sampling", imputation = impute_methods)
super_nhanes <- impute_missing(nhanes, "central_imputation")

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

#> Registered S3 method overwritten by 'quantmod':
#>   method            from
#>   as.zoo.data.frame zoo 

super_africa <- impute_missing(africa, "knn_imputation")
# Execute knn imputation with non default value for k
super_africa <- impute_missing(africa, "knn_imputation", k = 5)
super_africa <- impute_missing(africa, "expect_maximization", exclude = "country")
super_africa <- impute_missing(africa, "rf_imputation", num_iterations = 15,
                               num_trees = 200, bootstrap = FALSE)

#>   missForest iteration 1 in progress...done!
#>   missForest iteration 2 in progress...done!
#>   missForest iteration 3 in progress...done!
#>   missForest iteration 4 in progress...done!

# Examples of calls to 'PCA imputation' with wholly numeric datasets
# \donttest{
super_orange <- impute_missing(orange, "PCA_imputation", num_dimensions = 5,
                               imputation = "EM")

#> Warning: Stopped after  1000  iterations

super_orange <- impute_missing(orange, "PCA_imputation", num_dimensions = 5,
                               imputation = "Regularized")

#> Warning: Stopped after  1000  iterations

# }
super_orange <- impute_missing(orange, "PCA_imputation", num_dimensions = 5,
                               imputation = "Regularized", random_init = TRUE)

#> Warning: Stopped after  1000  iterations

# Examples of calls to 'MCA imputation' with wholly categorical datasets
# \donttest{
super_vnf    <- impute_missing(vnf, "MCA_imputation", num_dimensions = 5,
                               imputation = "EM")
super_vnf    <- impute_missing(vnf, "MCA_imputation", num_dimensions = 5,
                               imputation = "Regularized")
# }
super_vnf    <- impute_missing(vnf, "MCA_imputation", num_dimensions = 5,
                               imputation = "Regularized", random_init = TRUE)
# Examples of calls to 'FAMD imputation' with hybrid datasets
# \donttest{
super_ozone  <- impute_missing(ozone, "FAMD_imputation", num_dimensions = 5,
                               imputation = "EM", exclude = c("Ne12", "Vx15"))
super_ozone  <- impute_missing(ozone, "FAMD_imputation", num_dimensions = 5,
                               imputation = "Regularized")
# }
super_ozone  <- impute_missing(ozone, "FAMD_imputation", num_dimensions = 5,
                               imputation = "Regularized", random_init = TRUE)

# Examples of hotdeck, iterative robust and reggresion imputations
super_sleep <- impute_missing(sleep, "hotdeck")
super_sleep <- impute_missing(sleep, "iterative_robust", initialization = "median",
                              num_iterations = 1000)
super_sleep <- impute_missing(sleep, "regression_imputation",
                              formula = Dream+NonD~BodyWgt+BrainWgt)

# Examples of adaptative shrinkage imputation
super_ozone <- impute_missing(ozone, "ATN", sigma = 2.2)
super_ozone <- impute_missing(ozone, "ATN", lambda = 0.025, gamma = 2.5)
super_ozone <- impute_missing(ozone, "ATN", tune = "SURE")