smartdata: A package to ease data preprocessing tasks

Provides a pipe interface that integrates a collection of the most used data preprocessing libraries, providing oversampling, instance and feature selection, normalization, discretization, space transformation and outliers/noise/missing values treatment

Method to do oversampling oversample

NA

Method to do instance selection instance_selection

NA

Method to do feature selection feature_selection

NA

Method to do normalization normalize

NA

Method to do discretization discretize

NA

Method to do space transformation space_transformation

NA

Method to treat outliers clean_outliers

NA

Method to treat missing values impute_missing

NA

Method to treat noise clean_noise

NA