Prints options for a wrapper or a certain preprocessing method

Prints options for a wrapper or a certain preprocessing method

which_options(preprocess, method = NULL)

Arguments

preprocess
Possible preprocessing: 'oversample', 'clean_noise', 'instance_selection', 'feature_selection', 'normalize', 'discretize', 'space_transformation', 'clean_outliers', 'impute_missing', 'clean_noise'

method
For the preprocessing method

Value

Prints valid preprocessings for the selected wrapper or options for a given preprocessing

Examples

which_options("oversample")

#> Possible methods are: 'RACOG', 'wRACOG', 'PDFOS', 'RWO', 'ADASYN', 'ANSMOTE', 'SMOTE', 'MWMOTE', 'BLSMOTE', 'DBSMOTE', 'SLMOTE', 'RSLSMOTE' 

which_options("clean_noise", method = "edgeWeight")

#> For more information do: ?NoiseFiltersR::EWF 
#> Parameters for edgeWeight are: 
#>   * threshold: Number between 0 and 1. Limit between good and suspicious instances 
#>                Default value: 0.25 
#>   * action: Strategy to treat noisy instances: 'remove', 'hybrid' 
#>             Default value: remove 

which_options("clean_noise", method = "ENG")

#> For more information do: ?NoiseFiltersR::ENG 
#> Parameters for ENG are: 
#>   * graph: Character indicating the type of graph to be constructed. It can be 
#>            chosen between 'GG' (Gabriel Graph) and 'RNG' (Relative 
#>            Neighborhood Graph 
#>            Default value: GG 

which_options("impute_missing", method = "gibbs_sampling")

#> For more information do: ?mice::mice 
#> Parameters for gibbs_sampling are: 
#>   * imputation: Imputation method (one of the above or a vector containing as much as 
#>                 number of columns in dataset): 
#>                 pmm: Predictive mean matching 
#>                 midastouch: Weighted predictive mean matching 
#>                 sample: Random sample from observed values 
#>                 cart: Classification and regression trees 
#>                 rf: Random forest imputations 
#>                 mean: Unconditional mean imputation 
#>                 norm: Bayesian linear regression 
#>                 norm_nob: Linear regression ignoring model error 
#>                 norm_boot: Linear regression using bootstrap 
#>                 norm_predict: Linear regression, predicted values 
#>                 quadratic: Imputation of quadratic terms 
#>                 ri: Random indicator for nonignorable data 
#>                 logreg: Logistic regression 
#>                 logreg_boot: Logistic regression with bootstrap 
#>                 polr: Proportional odds model 
#>                 polyreg: Polytomous logistic regression 
#>                 lda: Linear discriminant analysis 
#>                 2l_norm: Level-1 normal heteroskedastic 
#>                 2l_lmer: Level-1 normal homoscedastic, lmer 
#>                 2l_pan: Level-1 normal homoscedastic, pan 
#>                 2lonly_mean: Level-2 class mean 
#>                 2lonly_norm: Level-2 class normal 
#>                 2lonly_pmm: Level-2 class predictive mean matching 
#>                 Default value: pmm 
#>   * num_iterations: Desired number of iterations in the Gibbs sampling 
#>                     Default value: 5